Use size_t for message length

+++ libgammu/gsmstate.h
@@ -1436,7 +1436,7 @@ struct _GSM_StateMachine {
 GSM_Error GSM_RegisterAllPhoneModules	(GSM_StateMachine *s);
 
 GSM_Error GSM_WaitForOnce		(GSM_StateMachine *s, unsigned const char *buffer,
-			  		 int length, int type, int timeout);
+			  		 size_t length, int type, int timeout);
 
 /**
  * Wait for reply from the phone.
@@ -1451,7 +1451,7 @@ GSM_Error GSM_WaitForOnce		(GSM_StateMachine *s, unsigned const char *buffer,
  * \return Error code, ERR_NONE on sucecss.
  */
 GSM_Error GSM_WaitFor			(GSM_StateMachine *s, unsigned const char *buffer,
-		       			 int length, int type, int timeout,
+		       			 size_t length, int type, int timeout,
 					 GSM_Phone_RequestID request) WARNUNUSED;
 
 /**
@@ -1472,10 +1472,10 @@ GSM_Error GSM_WaitFor			(GSM_StateMachine *s, unsigned const char *buffer,
 
 GSM_Error GSM_DispatchMessage		(GSM_StateMachine *s);
 
-void 	  GSM_DumpMessageText		(GSM_StateMachine *s, unsigned const char *message, int messagesize, int type);
-void 	  GSM_DumpMessageTextRecv	(GSM_StateMachine *s, unsigned const char *message, int messagesize, int type);
-void 	  GSM_DumpMessageBinary		(GSM_StateMachine *s, unsigned const char *message, int messagesize, int type);
-void GSM_DumpMessageBinaryRecv(GSM_StateMachine *s, unsigned const char *message, int messagesize, int type);
+void 	  GSM_DumpMessageText		(GSM_StateMachine *s, unsigned const char *message, size_t messagesize, int type);
+void 	  GSM_DumpMessageTextRecv	(GSM_StateMachine *s, unsigned const char *message, size_t messagesize, int type);
+void 	  GSM_DumpMessageBinary		(GSM_StateMachine *s, unsigned const char *message, size_t messagesize, int type);
+void GSM_DumpMessageBinaryRecv(GSM_StateMachine *s, unsigned const char *message, size_t messagesize, int type);
 
 
 void GSM_OSErrorInfo(GSM_StateMachine *s, const char *description);
